requirements.txt
Flask==2.2.3
openai==0.28.0
python-dotenv==0.21.0
python-docx
